requirements.txt
requests==2.25.1
XlsxWriter==1.3.7
python-dotenv==0.15.0
